+++ HISTORY.rst
@@ -8,6 +8,7 @@ Changelog
 - Basic extensions framework in place. TextBlob has been refactored to make it easier to develop extensions.
 - Add ``text.classifiers.PositiveNaiveBayesClassifier``.
 - Update NLTK.
+- Fix ``__str__`` behavior. ``print blob`` should now print expected output in both Python 2 and 3.
 - *Backwards-incompatible*: All abstract base classes have been moved to the ``text.base`` module.
 - *Backwards-incompatible*: ``PerceptronTagger`` will now be maintained as an extension, ``textblob-aptagger``. Instantiating a ``text.taggers.PerceptronTagger()`` will raise a ``DeprecationWarning``.